Rename image_convertion module to image_conversion
