+++ src/lib.rs
@@ -5,7 +5,7 @@ pub mod color_format;
 pub mod image;
 pub mod math;
 
-mod image_convertion;
+mod image_conversion;
 mod tiff_extentions;
